2to3

+++ plugins/dstat_freespace.py
@@ -35,7 +35,7 @@ class dstat_plugin(dstat):
         self.val['total'] = (0, 0)
         for name in self.vars:
             res = os.statvfs(name)
-            self.val[name] = ( (float(res.f_blocks) - float(res.f_bavail)) * long(res.f_frsize), float(res.f_bavail) * float(res.f_frsize) )
+            self.val[name] = ( (float(res.f_blocks) - float(res.f_bavail)) * int(res.f_frsize), float(res.f_bavail) * float(res.f_frsize) )
             self.val['total'] = (self.val['total'][0] + self.val[name][0], self.val['total'][1] + self.val[name][1])
 
 # vim:ts=4:sw=4:et
